GCListener: Use Environment.HasShutdownStarted to check for runtime shutdown
Using AppDomain.CurrentDomain.IsFinalizingForUnload() is not viable on Unity as the mono finalizer thread will be run before AppDomain is marked for unloading. This causes an infinite finalizer loop that locks mono runtime from shutting down. This fix checks both for root domain unloading and runtime shutdown at the same time which fixes the loop on mono.